Complete SQLite → Postgres migration: all 381 tests pass
Fixes for Neon Postgres compatibility:
- Pool config: max_connections(2), acquire_timeout(30s), DELETE FROM
  instead of TRUNCATE (avoids PgBouncer lock contention)
- Boolean columns: needs_enrichment = 0/1 → FALSE/TRUE
- Type mismatches: CAST(COUNT AS INTEGER) removed (native BIGINT),
  deezer_id i64→i32, generation_count read as i32
- Ambiguous ON CONFLICT: qualify column refs with table name
- Non-deterministic test: order-independent assertion for enrichment
- All migrations updated for Postgres dialect (DOUBLE PRECISION,
  TIMESTAMP, ON CONFLICT DO NOTHING, NOW())

+++ backend/migrations/003_dj_metadata.sql
@@ -1,8 +1,8 @@
 -- Migration 003: Add DJ metadata columns to tracks table
 -- Supports ST-001 (track catalog API) and UC-013 (multi-source import)
 
-ALTER TABLE tracks ADD COLUMN bpm REAL;
+ALTER TABLE tracks ADD COLUMN bpm DOUBLE PRECISION;
 ALTER TABLE tracks ADD COLUMN camelot_key TEXT;
-ALTER TABLE tracks ADD COLUMN energy REAL;
+ALTER TABLE tracks ADD COLUMN energy DOUBLE PRECISION;
 ALTER TABLE tracks ADD COLUMN source TEXT NOT NULL DEFAULT 'spotify';
 ALTER TABLE tracks ADD COLUMN album_art_url TEXT;
